# docs/specs/03-api/openapi.yaml
openapi: 3.1.0
info:
  title: AssoLink API
  version: 0.1.0
  description: |
    REST API d'AssoLink. Sécurité par session cookie HttpOnly + CSRF token.
    Voir ADR-0004 pour la stratégie d'authentification.
  contact:
    name: nic
  license:
    name: AGPL-3.0
servers:
  - url: https://api.local.hello-there.net/v1
    description: Production
  - url: https://staging-api.local.hello-there.net/v1
    description: Staging
  - url: http://localhost:8080/v1
    description: Dev local
security:
  - {}
...
paths:
  /auth/register/member:
    post:
      operationId: registerMember
      tags: [auth]
      summary: Inscrire une Personne
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterMember' }
      responses:
        '201': { description: Compte créé en pending_verification }
        '400': { $ref: '#/components/responses/ValidationError' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/register/association:
    post:
      operationId: registerAssociation
      tags: [auth]
      summary: Inscrire une Association
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterAssociation' }
      responses:
        '201': { description: Compte créé en pending_verification }
        '400': { $ref: '#/components/responses/ValidationError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id
  responses:
    ValidationError:
      description: Validation d'entrée échouée
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
              fields:
                type: object
                additionalProperties: { type: string }
    RateLimited:
      description: Trop de requêtes
      headers:
        Retry-After:
          schema: { type: integer }
  schemas:
    RegisterMember:
      type: object
      required: [email, password, first_name, last_name, birth_date, accept_tos]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12 }
        first_name: { type: string }
        last_name: { type: string }
        birth_date: { type: string, format: date }
        nickname: { type: string }
        accept_tos: { type: boolean }
    RegisterAssociation:
      type: object
      required: [email, password, legal_name, postal_code, accept_tos]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 12 }
        legal_name: { type: string }
        siren: { type: string }
        rna: { type: string }
        postal_code: { type: string }
        accept_tos: { type: boolean }
...